peliculas terminado

+++ ejercicios2/peliculas/index.php
@@ -8,36 +8,32 @@ $usuario = "root";
 $contrasena = "root";
 $db = "mydatabase";
 
-$conn = new mysqli($servidor, $usuario, $contrasena,$db);
+$conn = new mysqli($servidor, $usuario, $contrasena, $db);
 
-$usu = $_SESSION['nombre'];
+$usu = $_SESSION['usuario'];
 
 $sql1 = "SELECT * from peliculasUsuario where usuario = '$usu'";
-$result1 = $conn -> query($sql1);
+$result1 = $conn->query($sql1);
 
 
-
-
-
-if ($_SERVER["REQUEST_METHOD"] == "POST") {
+if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['submit'])) {
 
     if (isset($_POST['isan'])) {
 
         $isan = $_POST['isan'];
         $sql2 = "SELECT isan from peliculasUsuario where isan = '$isan'";
-        $result2 = $conn -> query($sql2);
+        $result2 = $conn->query($sql2);
         $res2 = $result2->fetch_assoc();
 
         if (isset($_POST['isan']) && empty($_POST['nombre'])) {
             $sql3 = "DELETE FROM peliculasUsuario where isan='$isan'";
-            $conn -> query($sql3);
+            $conn->query($sql3);
             $advertencia = "Pelicula borrada.";
-            
         }
-    
-    
+
+
         if (!empty($_POST['nombre']) && !empty($_POST['ano']) && !empty($_POST['puntuacion'])) {
-    
+
             $ano = $_POST['ano'];
             $puntuacion = $_POST['puntuacion'];
             $nombre = $_POST['nombre'];
@@ -45,40 +41,32 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
             if ($result2->num_rows > 0 && isset($_POST['nombre']) && isset($_POST['ano']) && isset($_POST['puntuacion'])) {
 
                 $sql4 = "UPDATE peliculasUsuario set ano = '$ano', puntuacion = '$puntuacion', nombre_pelicula = '$nombre' where isan = '$isan'";
-                $conn -> query($sql4);
+                $conn->query($sql4);
                 $advertencia = "Pelicula actualizada.";
-                
-                
-                
             }
-           
-        
-        
-            if ( !empty($_POST['ano']) && !empty($_POST['puntuacion'])) {
-                $sql5 = "INSERT INTO peliculasUsuario VALUES ('$usu','$isan','$nombre','$puntuacion','$ano')";
-                $conn -> query($sql5);
-                $advertencia = "Pelicula añadida.";
-                
-            }
-        
-        } 
 
+            if (!empty($_POST['ano']) && !empty($_POST['puntuacion'])) {
+                $sql5 = "INSERT INTO peliculasUsuario VALUES ('$usu','$isan','$nombre','$puntuacion','$ano')";
+                $conn->query($sql5);
+                $advertencia = "Pelicula añadida.";
+            }
+        }
     }
-    
 }
-while($row = $result1->fetch_assoc()) {
+
+while ($row = $result1->fetch_assoc()) {
 
     echo "Nombre: " . $row['nombre_pelicula'] . " / ISAN: " . $row['ISAN'] . " / Puntuacion: " . $row['puntuacion'] . " / Año: " . $row['ano'] . "<br>";
-
 }
 
 ?>
 
 <html>
-    <body>
+
+<body>
 
     <h1>Peliculas</h1>
-    <form method="post" action="<?php echo htmlspecialchars($_SERVER['PHP_SELF']);?>">
+    <form method="post" action="<?php echo htmlspecialchars($_SERVER['PHP_SELF']); ?>">
 
         Nombre: <input type="text" name="nombre">
         <br><br>
@@ -98,7 +86,10 @@ while($row = $result1->fetch_assoc()) {
         </select>
         <br><br>
         <input type="submit" name="submit" value="Submit">
-        <p><?php echo $advertencia?></p>
+        <form method="post" action="formularios.html">
+        <input type="submit" name="close" value="Cerrar sesion">
+        <p><?php echo $advertencia ?></p>
     </form>
-    </body>
+</body>
+
 </html>
